# ent/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ent Schema API",
    "description": "This is an auto generated API description made out of an Ent schema definition",
    "version": "0.1.0"
  },
  "paths": {
    "/endpoint-data": {
      "get": {
        "tags": [
          "EndpointData"
        ],
        "summary": "List EndpointData",
        "description": "List EndpointData.",
        "operationId": "listEndpointData",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "what page to render",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "item count to render per page",
            "schema": {
              "type": "integer",
              "maximum": 255,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "result EndpointData list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EndpointDataList"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "EndpointData"
        ],
        "summary": "Create a new EndpointData",
        "description": "Creates a new EndpointData and persists it to storage.",
        "operationId": "createEndpointData",
        "requestBody": {
          "description": "EndpointData to create",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dataName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "endpoint_required": {
                    "type": "integer"
                  },
                  "endpoint_provided": {
                    "type": "integer"
                  }
                },
                "required": [
                  "dataName",
                  "description",
                  "type"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "EndpointData created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointDataCreate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/endpoint-data/{id}": {
      "get": {
        "tags": [
          "EndpointData"
        ],
        "summary": "Find a EndpointData by ID",
        "description": "Finds the EndpointData with the requested ID and returns it.",
        "operationId": "readEndpointData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the EndpointData",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "EndpointData with requested ID was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointDataRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "EndpointData"
        ],
        "summary": "Deletes a EndpointData by ID",
        "description": "Deletes the EndpointData with the requested ID.",
        "operationId": "deleteEndpointData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the EndpointData",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "EndpointData with requested ID was deleted"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "tags": [
          "EndpointData"
        ],
        "summary": "Updates a EndpointData",
        "description": "Updates a EndpointData and persists changes to storage.",
        "operationId": "updateEndpointData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the EndpointData",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "EndpointData properties to update",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dataName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "endpoint_required": {
                    "type": "integer"
                  },
                  "endpoint_provided": {
                    "type": "integer"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "EndpointData updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointDataUpdate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/endpoint-data/{id}/endpoint-provided": {
      "get": {
        "tags": [
          "EndpointData"
        ],
        "summary": "Find the attached ProviderEndpoint",
        "description": "Find the attached ProviderEndpoint of the EndpointData with the given ID",
        "operationId": "readEndpointDataEndpointProvided",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the EndpointData",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ProviderEndpoint attached to EndpointData with requested ID was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointData_EndpointProvidedRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/endpoint-data/{id}/endpoint-required": {
      "get": {
        "tags": [
          "EndpointData"
        ],
        "summary": "Find the attached ProviderEndpoint",
        "description": "Find the attached ProviderEndpoint of the EndpointData with the given ID",
        "operationId": "readEndpointDataEndpointRequired",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the EndpointData",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ProviderEndpoint attached to EndpointData with requested ID was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointData_EndpointRequiredRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/provider-endpoints/{id}/provided-data": {
      "get": {
        "tags": [
          "ProviderEndpoint"
        ],
        "summary": "List attached ProvidedData",
        "description": "List attached ProvidedData.",
        "operationId": "listProviderEndpointProvidedData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the ProviderEndpoint",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "what page to render",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "item count to render per page",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "result ProviderEndpoints list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderEndpoint_ProvidedDataList"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/provider-endpoints/{id}/provider": {
      "get": {
        "tags": [
          "ProviderEndpoint"
        ],
        "summary": "Find the attached ProviderRegisterData",
        "description": "Find the attached ProviderRegisterData of the ProviderEndpoint with the given ID",
        "operationId": "readProviderEndpointProvider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the ProviderEndpoint",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ProviderRegisterData attached to ProviderEndpoint with requested ID was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEndpoint_ProviderRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EndpointDataList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "dataName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "dataName",
          "description",
          "type"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "invalid input, data invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
...
      "404": {
        "description": "resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
      "409": {
        "description": "conflicting resources",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
      "500": {
        "description": "unexpected error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      }
    }
  }
}